fit-vision-app/backend/requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary
supabase==2.23.2
python-dotenv==1.0.0
google-genai==1.51.0
redis==5.0.3
upstash-redis==1.0.0

# MCP Server dependencies
mcp>=1.1.0
PyGithub>=2.1.0
httpx>=0.27.0
# Development helpers
black==23.9.1
flake8==6.1.0

# Memory engine
tiktoken>=0.7.0
